+++ frontend/assets/entities/types.entity.ts
