+++ frontend/components/Common/MarkdownRenderers/Image.tsx
