+++ frontend/components/Configuration/defaultConfiguration.tsx
